requirements.txt
streamlit
pandas
sqlite3
numpy
plotly.express